move tuple, namedtuple, and typed dict to its own module
